"Flappy Bird"-like Game
Like the title said this is my final but the point of this code is a small "Flappy Bird"-like Game. Although I think majority of the code works and functions correctly I would love to add more stuff to it. This is a project I hope to grow for later.